requirements.txt
streamlit
google-generativeai
pyhwpx
Pillow